\name{calcDEMCnGenBurnin}
\alias{calcDEMCnGenBurnin}
\title{calcDEMCnGenBurnin}
\description{Number of steps based on expapolation of an observed temperature decrease to 1 	}
\usage{
calcDEMCnGenBurnin(T0, Ti, iStep)
}
\arguments{
  \item{T0}{the initial temperature (before the first step at iGen=0)}
  \item{Ti}{the temperatuer at step iStep}
  \item{iStep}{the step at which observed Ti}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





