\name{calcNGen.twDEMC}
\alias{calcNGen.twDEMC}
\title{calcNGen twDEMC}
\description{Calculates the number of completed generations in res}
\usage{
calcNGen.twDEMC(res, ...)
}
\arguments{
  \item{res}{object of class twDEMC}
  \item{\dots}{
}
}

\value{integer, number of completed generations}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCInt}}}

