\name{calcTemperatedLogDenChains}
\alias{calcTemperatedLogDenChains}
\title{calcTemperatedLogDenChains}
\description{Generic method behind \code{\link{calcTemperatedLogDenChains.array}}}
\usage{
calcTemperatedLogDenChains(...)
}
\arguments{
  \item{\dots}{
}
}



\author{Thomas Wutzler}





