\name{checkProblemsGelman1}
\alias{checkProblemsGelman1}
\title{checkProblemsGelman1}
\description{convergence between several populations}
\usage{checkProblemsGelman1(aTwDEMC, criticalGelmanDiag = 1.2)}
\arguments{
  \item{aTwDEMC}{
}
  \item{criticalGelmanDiag}{
}
}



\author{Thomas Wutzler}





