\name{checkProblemsSpectral}
\alias{checkProblemsSpectral}
\title{checkProblemsSpectral}
\description{Calculate and check problems on too high spectral ratios of Variance}
\usage{checkProblemsSpectral(aTwDEMC, critSpecVarRatio = 50)}
\arguments{
  \item{aTwDEMC}{twDEMC object to check}
  \item{critSpecVarRatio}{the threshold to report problems}
}



\author{Thomas Wutzler}





