\name{completeResCompVec}
\alias{completeResCompVec}
\title{completeResCompVec}
\description{given a vector with names, make a vector corresponding to resComp with components, not yet given filled by NA}
\usage{completeResCompVec(x, resCompNames)}\arguments{
  \item{x}{named vector to be completed}
  \item{resCompNames}{names of the result components}
}

\value{vector with names resCompNames, with corresponding values of x set, others NA}

\author{Thomas Wutzler}





