\name{denSparce}
\alias{denSparce}
\title{denSparce}
\description{Example of using two different logDensity functions: density of sparce observations}
\usage{
denSparce(theta,
twTwoDenEx = twTwoDenEx1,
theta0 = twTwoDenEx$thetaTrue,
...)
}
\arguments{
  \item{theta}{
}
  \item{twTwoDenEx}{
}
  \item{theta0}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler}





