\name{denSparcePrior}
\alias{denSparcePrior}
\title{denSparcePrior}
\description{Same as denSpace but returning two components: prior and misfit}
\usage{
denSparcePrior(theta,
twTwoDenEx = twTwoDenEx1,
theta0 = twTwoDenEx$thetaTrue,
thetaPrior = NULL,
invCovarTheta = NULL,
...)
}
\arguments{
  \item{theta}{
}
  \item{twTwoDenEx}{
}
  \item{theta0}{
}
  \item{thetaPrior}{the prior estimate of the parameters}
  \item{invCovarTheta}{the inverse of the Covariance of the prior parameter estimates}
  \item{\dots}{
}
}



\author{Thomas Wutzler}





