\name{denSparse}
\alias{denSparse}
\title{denSparse}
\description{Example of using two different logDensity functions: density of Sparse observations}
\usage{denSparse(theta, twTwoDenEx, theta0 = twTwoDenEx$thetaTrue, intermediate = NULL, 
    ...)}\arguments{
  \item{theta}{
}
  \item{twTwoDenEx}{
}
  \item{theta0}{
}
  \item{intermediate}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler}





