\name{divideTwDEMCBatch}
\alias{divideTwDEMCBatch}
\title{divideTwDEMCBatch}
\description{Generic method behind \code{\link{divideTwDEMCBatch.divideTwDEMCBatch}}}
\usage{
divideTwDEMCBatch(...)
}
\arguments{
  \item{\dots}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





