\name{dummyTwDEMCModel}
\alias{dummyTwDEMCModel}
\title{dummyTwDEMCModel}
\description{example model function: y=a+bx}
\usage{dummyTwDEMCModel(theta, xval)}
\arguments{
  \item{theta}{parameter vector with names a and b}
  \item{xval}{additional argument, passed by ... in logDenGaussian}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twDEMCBlockInt}}
\code{\link{logDenGaussian}}}

