\name{getCurrentTemp.twDEMC}
\alias{getCurrentTemp.twDEMC}
\title{getCurrentTemp twDEMC}
\description{Get the Temperature, i.e. cost reduction factor of the last sample}
\usage{\method{getCurrentTemp}{twDEMC}(x, ...)}
\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}
\details{There is only one sample per thinning interval of length \code{x$thin}.}


\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMC}}
\code{\link{subChains.twDEMC}}
\link{twDEMC}}
\examples{FALSE}
