\name{getCurrentTemp.twDEMCPops}
\alias{getCurrentTemp.twDEMCPops}
\title{getCurrentTemp twDEMCPops}
\description{Get the Temperature, i.e. cost reduction factor of the last sample}
\usage{\method{getCurrentTemp}{twDEMCPops}(x, ...)}\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}

