\name{getNBlocks.twDEMC}
\alias{getNBlocks.twDEMC}
\title{getNBlocks twDEMC}
\description{Extracts the number of blocks }
\usage{
\method{getNBlocks}{twDEMC}(x,
...)
}
\arguments{
  \item{x}{object of class twDEMC}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMC}}
,\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCBlockInt}}}

