\name{getNBlocks.twDEMCPops}
\alias{getNBlocks.twDEMCPops}
\title{getNBlocks twDEMCPops}
\description{Extracts the number of samples}
\usage{
\method{getNBlocks}{twDEMCPops}(x,
...)
}
\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}
\details{There is only one sample per thinning interval of length \code{x$thin}.}


\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}

