\name{getNChains.twDEMC}
\alias{getNChains.twDEMC}
\title{getNChains twDEMC}
\description{Extracts the number of chains}
\usage{\method{getNChains}{twDEMC}(res, ...)}\arguments{
  \item{res}{object of class twDEMC}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMC}}
\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCBlockInt}}}

