\name{getNChains.twDEMCPops}
\alias{getNChains.twDEMCPops}
\title{getNChains twDEMCPops}
\description{Extracts the number of chains}
\usage{
\method{getNChains}{twDEMCPops}(x,
...)
}
\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}

