\name{getNChainsPop.twDEMC}
\alias{getNChainsPop.twDEMC}
\title{getNChainsPop twDEMC}
\description{Extracts the number of chains per population}
\usage{\method{getNChainsPop}{twDEMC}(res, ...)}
\arguments{
  \item{res}{object of class twDEMC}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMC}}
\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCBlockInt}}}

