\name{getNChainsPop.twDEMCPops}
\alias{getNChainsPop.twDEMCPops}
\title{getNChainsPop twDEMCPops}
\description{Extracts the number of chains per population}
\usage{\method{getNChainsPop}{twDEMCPops}(x, ...)}
\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}

