\name{getNParms.twDEMC}
\alias{getNParms.twDEMC}
\title{getNParms twDEMC}
\description{Extracts the number of parameters, i.e. the dimension of the parameter vector}
\usage{\method{getNParms}{twDEMC}(res, ...)}\arguments{
  \item{res}{object of class twDEMC}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMC}}
\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCBlockInt}}}

