\name{getNParms.twDEMCPops}
\alias{getNParms.twDEMCPops}
\title{getNParms twDEMCPops}
\description{Extracts the number of parameters, i.e. the dimension of the parameter vector}
\usage{\method{getNParms}{twDEMCPops}(x, ...)}
\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}
\examples{FALSE}
