\name{getNPops.twDEMC}
\alias{getNPops.twDEMC}
\title{getNPops twDEMC}
\description{Extracts the number of populations}
\usage{
\method{getNPops}{twDEMC}(res,
...)
}
\arguments{
  \item{res}{object of class twDEMC}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMC}}
,\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCBlockInt}}}

