\name{getNPops.twDEMCPops}
\alias{getNPops.twDEMCPops}
\title{getNPops twDEMCPops}
\description{Extracts the number of populations}
\usage{\method{getNPops}{twDEMCPops}(x, ...)}
\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMCPops}}
,\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}
\examples{FALSE}
