\name{getNSamplesSpace.twDEMCPops}
\alias{getNSamplesSpace.twDEMCPops}
\title{getNSamplesSpace twDEMCPops}
\description{Extracts the number of samples summed over population within one space}
\usage{\method{getNSamplesSpace}{twDEMCPops}(x, ...)}
\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}
\details{There is only one sample per thinning interval of length \code{x$thin}.}


\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMCPops}}
\code{\link{getNSamples.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}
\examples{FALSE}
