\name{getNSpaces.twDEMCPops}
\alias{getNSpaces.twDEMCPops}
\title{getNSpaces twDEMCPops}
\description{Extracts the number of space replicates.}
\usage{\method{getNSpaces}{twDEMCPops}(x, ...)}\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getSpacesPop.twDEMCPops}}
\code{\link{getNGen.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}

