\name{getParBoundsPop.twDEMCPops}
\alias{getParBoundsPop.twDEMCPops}
\title{getParBoundsPop twDEMCPops}
\description{Extracts lower or upper ParBounds}
\usage{\method{getParBoundsPop}{twDEMCPops}(x, ...)}\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}

\value{
\item{upperParBoundsPop}{list of named numeric vectors giving upper parameter bounds per population}
\item{lowerParBoundsPop}{list of named numeric vectors giving lower parameter bounds per population
}}

\author{Thomas Wutzler}



\seealso{\code{\link{getNGen.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}

