\name{getSpacesPop.twDEMCPops}
\alias{getSpacesPop.twDEMCPops}
\title{getSpacesPop twDEMCPops}
\description{Extracts the space replicate that each population belongs to.}
\usage{
\method{getSpacesPop}{twDEMCPops}(x,
...)
}
\arguments{
  \item{x}{object of class twDEMCPops}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{getNSpaces.twDEMCPops}}
\code{\link{getNGen.twDEMCPops}}
\code{\link{subset.twDEMCPops}}
,\code{\link{twDEMCBlockInt}}}

