\name{logDenDs1MultiTemp}
\alias{logDenDs1MultiTemp}
\title{logDenDs1MultiTemp}
\description{Prior for theta}
\usage{
logDenDs1MultiTemp(theta, thetaPrior, sdTheta)
}
\arguments{
  \item{theta}{
}
  \item{thetaPrior}{
}
  \item{sdTheta}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





