\name{logLikDs1MultiTemp}
\alias{logLikDs1MultiTemp}
\title{logLikDs1MultiTemp}
\description{Prior for theta}
\usage{
logLikDs1MultiTemp(theta, thetaPrior, sdTheta)
}
\arguments{
  \item{theta}{
}
  \item{thetaPrior}{
}
  \item{sdTheta}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





