\name{orderLogDen}
\alias{orderLogDen}
\title{orderLogDen}
\description{rank the log-Densities, starting with highest densities}
\usage{orderLogDen(ss, nDen = 1)}
\arguments{
  \item{ss}{numeric matrix ( nRec x nDen+nPar )}
  \item{nDen}{number of densities}
}



\author{Thomas Wutzler}





