\name{plotConditional2D}
\alias{plotConditional2D}
\title{plotConditional2D}
\description{Plot the 2D conditional profile-Likelihood of a sample: calling \code{\link{plotMarginal2D}} with aggregating function max.}
\usage{
plotConditional2D(smp, xCol, yCol, intCol = 1, ...)
}
\arguments{
  \item{smp}{
}
  \item{xCol}{
}
  \item{yCol}{
}
  \item{intCol}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{plotMarginal2D}}
\code{\link{twDEMCInt}}}

