\name{replacePops.twDEMC}
\alias{replacePops.twDEMC}
\title{replacePops twDEMC}
\description{replaces several populations by toher ones}
\usage{
\method{replacePops}{twDEMC}(x, xNew, iPops, ...)
}
\arguments{
  \item{x}{
}
  \item{xNew}{
}
  \item{iPops}{
}
  \item{\dots}{
}
}

\value{a list of class twDEMC (see \code{\link{twDEMCInt}})}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCInt}}}

