\name{rinvchisq}
\alias{rinvchisq}
\title{rinvchisq}
\description{random number function and for the (scaled) inverse-chi-squared distribution.}
\usage{rinvchisq(n, df, scale = 1/df)}\arguments{
  \item{n}{
}
  \item{df}{
}
  \item{scale}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{rinvchisq}}}

