\name{sampleAcceptedFixedTempDiffLogDens}
\alias{sampleAcceptedFixedTempDiffLogDens}
\title{sampleAcceptedFixedTempDiffLogDens}
\description{Calculate diffLogDen for fixed temperature components and return subset of accepted ones }
\usage{
sampleAcceptedFixedTempDiffLogDens(diffLogDen, TFix = numeric(0))
}
\arguments{
  \item{diffLogDen}{array( streams x steps) Lp-La see \code{\link{getDiffLogDen.twDEMCProps}}}
  \item{TFix}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





