\name{sampleAcceptedFixedTempDiffLogliks}
\alias{sampleAcceptedFixedTempDiffLogliks}
\title{sampleAcceptedFixedTempDiffLogliks}
\description{Calculate diffLogLik for fixed temperature components and return subset of accepted ones }
\usage{
sampleAcceptedFixedTempDiffLogliks(diffLogLik, TFix = numeric(0))
}
\arguments{
  \item{diffLogLik}{array( streams x steps) Lp-La see \code{\link{getDiffLogLik.twDEMCProps}}}
  \item{TFix}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





