\name{stackChains.array}
\alias{stackChains.array}
\title{stackChains array}
\description{Combine MarkovChains of a initializer of twDEMC. }
\usage{
stackChains.array(Zinit, ...)
}
\arguments{
  \item{Zinit}{
}
  \item{\dots}{
}
}

\value{Matrix with columns the variables.}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{subChains.twDEMC}}}

