\name{stackChains.mcmc.list}
\alias{stackChains.mcmc.list}
\title{stackChains mcmc list}
\description{Combine list of MarkovChains to one big matrix.}
\usage{\method{stackChains}{mcmc.list}(mcmcList, ...)}\arguments{
  \item{mcmcList}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{subChains.twDEMC}}}

