\name{stackChains.twDEMC}
\alias{stackChains.twDEMC}
\title{stackChains twDEMC}
\description{Combine MarkovChains of a twDEMC to a matrix. }
\usage{
\method{stackChains}{twDEMC}(x, ...)
}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}

\value{Matrix with first column the log-Likelihold rLogLik and the remaining columns the variables.}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{subChains.twDEMC}}}

