\name{stackChains.twDEMCPops}
\alias{stackChains.twDEMCPops}
\title{stackChains twDEMCPops}
\description{Combine logLik and parms of MarkovChains of a twDEMCPops object to one matrix. }
\usage{\method{stackChains}{twDEMCPops}(x, ...)}
\arguments{
  \item{x}{
}
  \item{\dots}{not used}
}



\author{Thomas Wutzler}



\seealso{\code{\link{stackPopsInSpace}}, \code{\link{subset.twDEMCPops}} }

