\name{stackPops.twDEMCPops}
\alias{stackPops.twDEMCPops}
\title{stackPops twDEMCPops}
\description{Combine all populations (across subspaces) to one big population }
\usage{\method{stackPops}{twDEMCPops}(x, ...)}
\arguments{
  \item{x}{
}
  \item{\dots}{arguments passed \code{\link{combineTwDEMCPops}} such as \code{mergeMethod=stack/slice/random}}
}



\author{Thomas Wutzler}



\seealso{\code{\link{stackPopsInSpace.twDEMCPops}}
\code{\link{combineTwDEMCPops}}
\code{\link{stackChains.twDEMCPops}}
\code{\link{subset.twDEMCPops}}}
\examples{FALSE}
