\name{stackPopsInSpace.twDEMCPops}
\alias{stackPopsInSpace.twDEMCPops}
\title{stackPopsInSpace twDEMCPops}
\description{Combine populations for subspaces to bigger populations }
\usage{\method{stackPopsInSpace}{twDEMCPops}(x, ..., spacePop = getSpacesPop(x))}\arguments{
  \item{x}{
}
  \item{\dots}{arguments passed \code{\link{combineTwDEMCPops}} such as \code{mergeMethod=stack/slice/random}}
  \item{spacePop}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{stackPops.twDEMCPops}}
\code{\link{combineTwDEMCPops}}
\code{\link{stackChains.twDEMCPops}}
\code{\link{subset.twDEMCPops}}}

