\name{stat_prior}
\alias{stat_prior}
\title{stat prior}
\description{constructs a new StatPrior statistics based on aesthetics x and parName}
\usage{
stat_prior(mapping = NULL, data = NULL, geom = "line", position = "stack", adjust = 1, poptDistr, doTransOrig = FALSE, ...)
}
\arguments{
  \item{mapping}{
}
  \item{data}{
}
  \item{geom}{
}
  \item{position}{
}
  \item{adjust}{
}
  \item{poptDistr}{
}
  \item{doTransOrig}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





