\name{subPops.twDEMCPops}
\alias{subPops.twDEMCPops}
\title{subPops twDEMCPops}
\description{Condenses an twDEMCPops List to given population indices or space replicates}
\usage{
\method{subPops}{twDEMCPops}(x,
iPops,
iSpaces = NULL,
...)
}
\arguments{
  \item{x}{
}
  \item{iPops}{populations to keep}
  \item{iSpaces}{alternatively specifying the space replicates for which to keep populations}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twDEMCBlockInt}}
\code{\link{subset.twDEMCPops}}}

