\name{sumLogDenCompBlocks}
\alias{sumLogDenCompBlocks}
\title{sumLogDenCompBlocks}
\description{sum logDensity components across blocks}
\usage{sumLogDenCompBlocks(resLogDen, dInfos)}\arguments{
  \item{resLogDen}{numeric matrix (nStep x nResComp): logDensity (highest are best)}
  \item{dInfos}{list of lists with entry resCompPos (integer vector) specifying the position of result components for each density}
}



\author{Thomas Wutzler}



\seealso{\code{\link{orderLogDen}}, \code{\link{calcTemperatedLogDen}}}

