\name{twDEMC-internal}
\title{Internal twDEMC objects}
\alias{.calcRhat2}
\alias{.calcTemperatedDiffLogLik}
\alias{.checkConvergenceGelmanPops}
\alias{.detrendMatrix}
\alias{.doDEMCStep}
\alias{.doDEMCSteps}
\alias{.generateXPropThin}
\alias{.getResFLogLikNames}
\alias{.ofLnormMLE}
\alias{.tmp.f}
\alias{.tmp.f.plot.calcDEMCTempDiffLogLik2}
\alias{.twCoefLnormMLE_depr}
\alias{.xStepParallel}
\alias{.xStepSnooker}
\description{Internal twDEMC objects.}
\details{These are not to be called by the user.}
\keyword{internal}