\name{twDEMC.twDEMC}
\alias{twDEMC.twDEMC}
\title{twDEMC twDEMC}
\description{initialize \code{\link{twDEMCInt}} by former run and append results to former run}
\usage{
\method{twDEMC}{twDEMC}(Zinit, ...)
}
\arguments{
  \item{Zinit}{list of class twDEMC, result of \code{\link{twDEMCInt}}}
  \item{\dots}{further arguments to \code{\link{twDEMCInt}}}
}

\value{Zinit appended with the further generations of \code{\link{twDEMCInt}} }

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





