\name{twDEMCBlock.twDEMC}
\alias{twDEMCBlock.twDEMC}
\title{twDEMCBlock twDEMC}
\description{initialize \code{\link{twDEMCBlockInt}} by former run and append results to former run}
\usage{\method{twDEMCBlock}{twDEMC}(x, ...)}\arguments{
  \item{x}{list of class twDEMC, result of \code{\link{twDEMCBlockInt}}}
  \item{\dots}{further arguments to \code{\link{twDEMCBlockInt}}}
}



\author{Thomas Wutzler}





