\name{twDEMCPopMeans}
\alias{twDEMCPopMeans}
\title{twDEMCPopMeans}
\description{Calculating population means across chains, and smooth time series.}
\usage{
twDEMCPopMeans(x, nPops, kSmooth = NULL)
}
\arguments{
  \item{x}{a matrix with columns chains}
  \item{nPops}{number of populations}
  \item{kSmooth}{weights to the filter function, or just number of points}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





