\name{twVarDistrVec}
\alias{twVarDistrVec}
\title{twVarDistrVec}
\description{create a factor level with given variable names}
\usage{twVarDistrVec(varNames)}\arguments{
  \item{varNames}{character vector}
}

\value{a named factorvector initialized by level "norm"}

\author{Thomas Wutzler}





