\name{calcTemperatedLogDenChains}
\alias{calcTemperatedLogDenChains}
\title{calcTemperatedLogDenChains}
\description{Generic method behind \code{\link{calcTemperatedLogDenChains.twDEMC}}}
\usage{calcTemperatedLogDenChains(...)}
\arguments{
  \item{\dots}{
}
}



\author{Thomas Wutzler}





