\name{calcTemperatedLogDenChains.twDEMC}
\alias{calcTemperatedLogDenChains.twDEMC}
\title{calcTemperatedLogDenChains twDEMC}
\description{Rescale Log-Densities by given Temperatures within chains}
\usage{\method{calcTemperatedLogDenChains}{twDEMC}(x, temp = getCurrentTemp(x), 
    ...)}
\arguments{
  \item{x}{object of class twDEMC}
  \item{temp}{numeric vector (nResComp): Temperature, i.e. cost reduction factor}
  \item{\dots}{
}
}



\author{Thomas Wutzler}





