\name{denRich}
\alias{denRich}
\title{denRich}
\description{Example of using two different logDensity functions: density of data-rich observations}
\usage{
denRich(theta,
twTwoDenEx = twTwoDenEx1,
theta0 = twTwoDenEx$thetaTrue,
...)
}
\arguments{
  \item{theta}{
}
  \item{twTwoDenEx}{
}
  \item{theta0}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler}





