\name{denRich}
\alias{denRich}
\title{denRich}
\description{Example of using two different logDensity functions: density of data-rich observations}
\usage{denRich(theta, twTwoDenEx, theta0 = twTwoDenEx$thetaTrue, modTwoDenExCache = NULL, 
    ...)}
\arguments{
  \item{theta}{
}
  \item{twTwoDenEx}{
}
  \item{theta0}{
}
  \item{modTwoDenExCache}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler}





