\name{dummyTwDEMCModel}
\alias{dummyTwDEMCModel}
\title{dummyTwDEMCModel}
\description{example model function: y=a+bx}
\usage{
dummyTwDEMCModel(theta, xval)
}
\arguments{
  \item{theta}{parameter vector with names a and b}
  \item{xval}{additional argument, passed by ... in logLikGaussian}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{twDEMCInt}}
\code{\link{logLikGaussian}}}

