\name{mcmcListApply}
\alias{mcmcListApply}
\title{mcmcListApply}
\description{Applies function to each chain of \code{mcmc.list}.}
\usage{
mcmcListApply(mcmcList, fun, ...)
}
\arguments{
  \item{mcmcList}{
}
  \item{fun}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{subChains.twDEMC}}}

