\name{stackChains.array}
\alias{stackChains.array}
\title{stackChains array}
\description{Combine MarkovChains of result population of twDEMC. }
\usage{\method{stackChains}{array}(Zinit, ...)}\arguments{
  \item{Zinit}{
}
  \item{\dots}{
}
}

\value{Numeric matrix (nStep x nParm). steps=c(steps_chain1, steps_chain2_, ... )}

\author{Thomas Wutzler}



\seealso{\code{\link{subChains.twDEMC}}}

