\name{stackChains.twDEMCPops}
\alias{stackChains.twDEMCPops}
\title{stackChains twDEMCPops}
\description{Combine logLik and parms of MarkovChains of a twDEMCPops object to one matrix. }
\usage{\method{stackChains}{twDEMCPops}(x,
 ..., useTemperatedLogDen = FALSE, returnComponents = FALSE)}\arguments{
  \item{x}{
}
  \item{\dots}{not used}
  \item{useTemperatedLogDen}{set to true to calculate temperatedLogDen}
  \item{returnComponents}{set to TRUE to return logDenComp instead of logDen}
}



\author{Thomas Wutzler}



\seealso{\code{\link{stackPopsInSpace.twDEMCPops}}, \code{\link{subset.twDEMCPops}} }

